requirements.txt
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.7
django-allauth==0.24.1
django-contrib-comments==1.6.2
django-location-field==1.6.1
django-redis==4.3.0
django-toolbelt==0.0.1
django-watson==1.2.0
django-widget-tweaks==1.4.1
ExifRead==2.1.2
fake-factory==0.5.3
gunicorn==19.4.5
oauthlib==1.0.3
Pillow==3.1.0
psycopg2==2.6.1
python3-openid==3.0.9
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.6.0
six==1.10.0
sorl-thumbnail==12.3
static3==0.6.1
wheel==0.24.0
